# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Dog List API",
        "description": "API para gerenciamento de usuários e cachorros",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/dogs/store": {
            "post": {
                "tags": [
                    "Cachorros"
                ],
                "summary": "Adiciona um novo cachorro no banco de dados [requer JWT].",
                "operationId": "f7369c128c2be35a8f9cc293310712da",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name",
                                    "breed",
                                    "gender",
                                    "is_public"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Defina o nome de seu cachorro.",
                                        "type": "string",
                                        "example": "Buddy"
                                    },
                                    "breed": {
                                        "description": "Defina a raça de seu cachorro.",
                                        "type": "string",
                                        "example": "Golden Retriever"
                                    },
                                    "gender": {
                                        "description": "Defina o sexo do cachorro: M para masculino e F para feminino.",
                                        "type": "string",
                                        "enum": [
                                            "M",
                                            "F"
                                        ],
                                        "example": "M"
                                    },
                                    "is_public": {
                                        "description": "Defina 1 para tornar o cachorro público e 0 para mantê-lo privado.",
                                        "type": "integer",
                                        "enum": [
                                            0,
                                            1
                                        ],
                                        "example": 1
                                    },
                                    "img_path": {
                                        "description": "Nova imagem do cachorro",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Cachorro adicionado com sucesso"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/dogs/current_dog/{id}": {
            "get": {
                "tags": [
                    "Cachorros"
                ],
                "summary": "Obtem detalhes de um cachorro específico desde que esteja atrelado ao usuário logado [requer JWT].",
                "operationId": "2fef374324c47c9385175b6d83dfc572",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do cachorro",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalhes do cachorro retornados com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Buddy"
                                        },
                                        "breed": {
                                            "type": "string",
                                            "example": "Golden Retriever"
                                        },
                                        "gender": {
                                            "type": "string",
                                            "example": "M"
                                        },
                                        "is_public": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "img_path": {
                                            "type": "string",
                                            "example": "images/user/1/dog.jpg"
                                        },
                                        "user_id": {
                                            "type": "integer",
                                            "example": 1
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Acesso negado: o cachorro pertence a outro usuário"
                    },
                    "404": {
                        "description": "Cachorro não encontrado"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/dogs/update/{id}": {
            "post": {
                "tags": [
                    "Cachorros"
                ],
                "summary": "Edita informações de seu cachorro. Para isso é necessário informa seu id e preencher os campos do formulário [requer JWT].",
                "description": "Esta rota aceita um método POST com o campo `_method` definido como PUT para atualizar informações do cachorro.",
                "operationId": "9bb7611491878d42fe4b03cebd076ea7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do cachorro a ser atualizado",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "_method",
                                    "name",
                                    "breed",
                                    "gender",
                                    "is_public"
                                ],
                                "properties": {
                                    "_method": {
                                        "description": "Informa o método HTTP desejado. Mantenha o valor como PUT para que a atualização seja realizada corretamente.",
                                        "type": "string",
                                        "example": "PUT"
                                    },
                                    "name": {
                                        "description": "Defina o nome de seu cachorro.",
                                        "type": "string",
                                        "example": "Buddy"
                                    },
                                    "breed": {
                                        "description": "Defina a raça de seu cachorro.",
                                        "type": "string",
                                        "example": "Golden Retriever"
                                    },
                                    "gender": {
                                        "description": "Defina o sexo do cachorro: M para masculino e F para feminino.",
                                        "type": "string",
                                        "enum": [
                                            "M",
                                            "F"
                                        ],
                                        "example": "M"
                                    },
                                    "is_public": {
                                        "description": "Defina 1 para tornar o cachorro público e 0 para mantê-lo privado.",
                                        "type": "integer",
                                        "enum": [
                                            0,
                                            1
                                        ],
                                        "example": 1
                                    },
                                    "img_path": {
                                        "description": "Nova imagem do cachorro",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Cachorro atualizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Cachorro atualizado com sucesso"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro nos dados enviados"
                    },
                    "401": {
                        "description": "Token JWT inválido ou ausente"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/dogs/delete/{id}": {
            "delete": {
                "tags": [
                    "Cachorros"
                ],
                "summary": "Exclui seu cachorro do banco dados, basta informa seu id [requer JWT].",
                "operationId": "736db1e592a9acae7673ad9f3f2b11d3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do cachorro",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cachorro excluído com sucesso"
                    },
                    "403": {
                        "description": "Acesso negado: o cachorro pertence a outro usuário"
                    },
                    "404": {
                        "description": "Cachorro não encontrado"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/dogs/delete_image/{id}": {
            "put": {
                "tags": [
                    "Cachorros"
                ],
                "summary": "Apaga a imagem de seu cachorro, basta informar seu id [requer JWT].",
                "operationId": "d7aed838f6c2fb54ec9c60b8197b767b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do cachorro",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Imagem excluída com sucesso"
                    },
                    "403": {
                        "description": "Acesso negado: o cachorro pertence a outro usuário"
                    },
                    "404": {
                        "description": "Cachorro não encontrado"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearer": {
                "type": "http",
                "description": "Insira o token JWT no formato: Bearer {seu_token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
    "security": [
        {
            "bearer": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    ]
}